MathEx.sum(int[]) returns long

+++ core/src/main/java/smile/base/cart/DecisionNode.java
@@ -45,7 +45,7 @@ public class DecisionNode extends LeafNode {
      * @param count the number of node samples in each class.
      */
     public DecisionNode(int[] count) {
-        super(MathEx.sum(count));
+        super((int) MathEx.sum(count));
         this.output = MathEx.whichMax(count);
         this.count = count;
     }
